Multi-tenency. v.9

+++ src/main/java/org/servantscode/schedule/db/EquipmentDB.java
@@ -6,6 +6,7 @@ import org.servantscode.commons.AutoCompleteComparator;
 import org.servantscode.commons.db.DBAccess;
 import org.servantscode.commons.search.QueryBuilder;
 import org.servantscode.commons.search.SearchParser;
+import org.servantscode.commons.security.OrganizationContext;
 import org.servantscode.schedule.Equipment;
 
 import java.sql.*;
@@ -26,7 +27,7 @@ public class EquipmentDB extends DBAccess {
     }
 
     public int getCount(String search) {
-        QueryBuilder query = count().from("equipment").search(searchParser.parse(search));
+        QueryBuilder query = count().from("equipment").search(searchParser.parse(search)).inOrg();
         try (Connection conn = getConnection();
              PreparedStatement stmt = query.prepareStatement(conn);
              ResultSet rs = stmt.executeQuery()) {
@@ -40,7 +41,7 @@ public class EquipmentDB extends DBAccess {
     }
 
     public Equipment getEquipment(int id) {
-        QueryBuilder query = selectAll().from("equipment").withId(id);
+        QueryBuilder query = selectAll().from("equipment").withId(id).inOrg();
         try (Connection conn = getConnection();
              PreparedStatement stmt = query.prepareStatement(conn);
         ) {
@@ -53,7 +54,7 @@ public class EquipmentDB extends DBAccess {
     }
 
     public List<Equipment> getEquipmentList(String search, String sortField, int start, int count) {
-        QueryBuilder query = selectAll().from("equipment").search(searchParser.parse(search))
+        QueryBuilder query = selectAll().from("equipment").search(searchParser.parse(search)).inOrg()
                 .sort(sortField).limit(count).offset(start);
         try ( Connection conn = getConnection();
               PreparedStatement stmt = query.prepareStatement(conn)
@@ -66,12 +67,13 @@ public class EquipmentDB extends DBAccess {
 
     public Equipment create(Equipment equipment) {
         try (Connection conn = getConnection();
-             PreparedStatement stmt = conn.prepareStatement("INSERT INTO equipment(name, manufacturer, description) values (?, ?, ?)", Statement.RETURN_GENERATED_KEYS)
+             PreparedStatement stmt = conn.prepareStatement("INSERT INTO equipment(name, manufacturer, description, org_id) values (?, ?, ?, ?)", Statement.RETURN_GENERATED_KEYS)
         ){
 
             stmt.setString(1, equipment.getName());
             stmt.setString(2, equipment.getManufacturer());
             stmt.setString(3, equipment.getDescription());
+            stmt.setInt(4, OrganizationContext.orgId());
 
             if(stmt.executeUpdate() == 0) {
                 throw new RuntimeException("Could not create equipment: " + equipment.getName());
@@ -89,13 +91,14 @@ public class EquipmentDB extends DBAccess {
 
     public Equipment updateEquipment(Equipment equipment) {
         try (Connection conn = getConnection();
-             PreparedStatement stmt = conn.prepareStatement("UPDATE equipment SET name=?, manufacturer=?, description=? WHERE id=?")
+             PreparedStatement stmt = conn.prepareStatement("UPDATE equipment SET name=?, manufacturer=?, description=? WHERE id=? AND org_id=?")
         ) {
 
             stmt.setString(1, equipment.getName());
             stmt.setString(2, equipment.getManufacturer());
             stmt.setString(3, equipment.getDescription());
             stmt.setInt(4, equipment.getId());
+            stmt.setInt(5, OrganizationContext.orgId());
 
             if (stmt.executeUpdate() == 0)
                 throw new RuntimeException("Could not update equipment: " + equipment.getName());
@@ -108,10 +111,11 @@ public class EquipmentDB extends DBAccess {
 
     public boolean deleteEquipment(int id) {
         try (Connection conn = getConnection();
-             PreparedStatement stmt = conn.prepareStatement("DELETE FROM equipment WHERE id=?")
+             PreparedStatement stmt = conn.prepareStatement("DELETE FROM equipment WHERE id=? AND org_id=?")
         ) {
 
             stmt.setInt(1, id);
+            stmt.setInt(2, OrganizationContext.orgId());
             return stmt.executeUpdate() != 0;
         } catch (SQLException e) {
             throw new RuntimeException("Could not delete equipment: " + id, e);
